+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # igraph Python interface changelog
 
+## [master]
+
+### Added 
+
+- Added `is_bigraphical()` to test whether a pair of integer sequences can be the degree sequence of some bipartite graph.
+
 ## [0.10.7] - 2023-09-04
 
 ### Added